backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
numpy>=1.26.0
sentence-transformers>=2.6.1
pinecone>=3.0.0
python-dotenv>=1.0.1
pydantic>=2.6.0
typing-extensions>=4.12.0
